Split in-memory transport module into several smaller ones
