+++ src/test_fixture/network/mod.rs
@@ -1,3 +1,5 @@
+mod receive;
+mod stream;
 mod transport;
 
 use crate::{
