+++ src/test_fixture/network/receive.rs
@@ -1,0 +1,57 @@
+use crate::test_fixture::network::stream::{StreamCollection, StreamKey};
+use futures::Stream;
+use futures_util::StreamExt;
+use std::{
+    pin::Pin,
+    task::{Context, Poll},
+};
+
+/// Represents a stream of records.
+/// If stream is not received yet, each poll generates a waker that is used internally to wake up
+/// the task when stream is received.
+/// Once stream is received, it is moved to this struct and it acts as a proxy to it.
+pub struct ReceiveRecords<S> {
+    inner: ReceiveRecordsInner<S>,
+}
+
+impl<S> ReceiveRecords<S> {
+    pub(crate) fn new(key: StreamKey, coll: StreamCollection<S>) -> Self {
+        Self {
+            inner: ReceiveRecordsInner::Pending(key, coll),
+        }
+    }
+}
+
+impl<S: Stream + Unpin> Stream for ReceiveRecords<S> {
+    type Item = S::Item;
+
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        Pin::get_mut(self).inner.poll_next_unpin(cx)
+    }
+}
+
+/// Inner state for [`ReceiveRecords`] struct
+enum ReceiveRecordsInner<S> {
+    Pending(StreamKey, StreamCollection<S>),
+    Ready(S),
+}
+
+impl<S: Stream + Unpin> Stream for ReceiveRecordsInner<S> {
+    type Item = S::Item;
+
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        let this = Pin::get_mut(self);
+        loop {
+            match this {
+                Self::Pending(key, streams) => {
+                    if let Some(stream) = streams.add_waker(key, cx.waker()) {
+                        *this = Self::Ready(stream);
+                    } else {
+                        return Poll::Pending;
+                    }
+                }
+                Self::Ready(stream) => return stream.poll_next_unpin(cx),
+            }
+        }
+    }
+}
